feat(hotels): create rooms and hotels models and services
